Clean up front page
