--- polaris.shopify.com/src/components/ResourcesPage/index.ts
@@ -1,3 +1,0 @@
-import ResourcesPage from "./ResourcesPage";
-
-export default ResourcesPage;
