+++ polaris.shopify.com/src/data/foundationsIcons.tsx
@@ -103,6 +103,21 @@ export const InformationArchitectureIcon = (
   </svg>
 );
 
+export const CurrencyIcon = (
+  <svg
+    width="20"
+    height="20"
+    viewBox="0 0 20 20"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      d="M20 10C20 4.486 15.514 0 10 0C4.486 0 0 4.486 0 10C0 15.514 4.486 20 10 20C15.514 20 20 15.514 20 10ZM10.975 11.754H12.225C12.3235 11.754 12.421 11.7734 12.512 11.8111C12.603 11.8488 12.6857 11.904 12.7553 11.9737C12.825 12.0433 12.8802 12.126 12.9179 12.217C12.9556 12.308 12.975 12.4055 12.975 12.504C12.975 12.6025 12.9556 12.7 12.9179 12.791C12.8802 12.882 12.825 12.9647 12.7553 13.0343C12.6857 13.104 12.603 13.1592 12.512 13.1969C12.421 13.2346 12.3235 13.254 12.225 13.254H10.975V14.254C10.975 14.5192 10.8696 14.7736 10.6821 14.9611C10.4946 15.1486 10.2402 15.254 9.975 15.254C9.70978 15.254 9.45543 15.1486 9.26789 14.9611C9.08036 14.7736 8.975 14.5192 8.975 14.254V13.254H7.725C7.62651 13.254 7.52898 13.2346 7.43799 13.1969C7.34699 13.1592 7.26431 13.104 7.19467 13.0343C7.12503 12.9647 7.06978 12.882 7.03209 12.791C6.9944 12.7 6.975 12.6025 6.975 12.504C6.975 12.4055 6.9944 12.308 7.03209 12.217C7.06978 12.126 7.12503 12.0433 7.19467 11.9737C7.26431 11.904 7.34699 11.8488 7.43799 11.8111C7.52898 11.7734 7.62651 11.754 7.725 11.754H8.975V11.004H7.725C7.52609 11.004 7.33532 10.925 7.19467 10.7843C7.05402 10.6437 6.975 10.4529 6.975 10.254C6.975 10.0551 7.05402 9.86432 7.19467 9.72367C7.33532 9.58302 7.52609 9.504 7.725 9.504H8.584L6.734 6.3C6.60348 6.07044 6.56906 5.79858 6.63825 5.54374C6.70744 5.28889 6.87462 5.07176 7.10331 4.93973C7.33199 4.80769 7.60362 4.77147 7.85892 4.83896C8.11421 4.90646 8.33245 5.07219 8.466 5.3L9.981 7.924L11.496 5.3C11.6296 5.07219 11.8478 4.90646 12.1031 4.83896C12.3584 4.77147 12.63 4.80769 12.8587 4.93973C13.0874 5.07176 13.2546 5.28889 13.3237 5.54374C13.3929 5.79858 13.3585 6.07044 13.228 6.3L11.378 9.504H12.225C12.4239 9.504 12.6147 9.58302 12.7553 9.72367C12.896 9.86432 12.975 10.0551 12.975 10.254C12.975 10.4529 12.896 10.6437 12.7553 10.7843C12.6147 10.925 12.4239 11.004 12.225 11.004H10.975V11.754Z"
+      fill="#5C5F62"
+    />
+  </svg>
+);
+
 export const DesignIcon = (
   <svg
     width="20"
